refactor: split mapview component
